ListGlobalStyle•reset.js move route
